+++ lib/shipcloud/webhook.rb
@@ -1,0 +1,7 @@
+module Shipcloud
+  class Webhook < Base
+    include Shipcloud::Operations::All
+
+    attr_reader :url, :event_types
+  end
+end