+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - This CHANGELOG file to hopefully serve as an evolving example of a standardized open source project CHANGELOG.
 - Create, find, update and index operations for address resource. (#4)
 - Services attribute to carriers call (#5)
+- Create, find and index operations for webhook resource. (#7)
 - Added the following ruby versions to travis-ci test runs:
   - 2.1.7
   - 2.2.4